DBZ-6068 Remove deprecated delete/tombstone config options
